fix: usage of input for reservations
