+++ src/app/home/home.component.html
@@ -2,7 +2,7 @@
   <h1 class="w-min whitespace-nowrap text-3xl font-bold">Welkom, {{ this.user.firstName }}</h1>
   <div class="flex flex-col w-full space-y-5 md:space-y-0 md:space-x-20 md:flex-row">
     <section class="w-full md:w-3/5">
-      <app-upcoming-reservations></app-upcoming-reservations>
+      <app-upcoming-reservations [reservations]="reservations" ></app-upcoming-reservations>
       <app-default-location [defaultLocation]="favoriteLocation"></app-default-location>
     </section>
