+++ src/app/home/upcoming-reservations/upcoming-reservations.component.ts
@@ -17,23 +17,8 @@ import { Reservation } from '../../shared/model/reservation.model';
   styleUrl: './upcoming-reservations.component.scss'
 })
 export class UpcomingReservationsComponent {
+  @Input()
   reservations!: Reservation[];
 
   constructor(private reservationService: ReservationService) { }
-
-  ngOnInit(): void {
-    this.getUpcomingReservations();
-  }
-
-  async getUpcomingReservations(): Promise<void> {
-    let allReservations = await this.reservationService.getAllReservations();
-    let now = new Date();
-
-    let upcomingReservations = allReservations.filter((reservation) => {
-      let startDateTime = new Date(reservation.startDateTime);
-      return startDateTime >= now;
-    });
-    
-    this.reservations = upcomingReservations.sort((a, b) => new Date(a.startDateTime).getTime() - new Date(b.startDateTime).getTime()).slice(0, 3);
-  }
 }
